Downloads/spotify git/requirements.txt
spotipy==2.22.1
Flask==2.3.3
RPi.GPIO==0.7.1
requests==2.31.0
psutil==5.9.5
python-dotenv==1.0.0